re-added os implementation for windows split windows and linux data code into their own class files for sanity dev logging mostly disabled prompt log now works with split os classes
